feat: add delete establishment

+++ src/repositories/establishment-prisma-repository.ts
@@ -5,6 +5,9 @@ import { Establishment, Prisma } from "@prisma/client";
 export class EstablishmentPrismaRepository implements IEstablishmentRepository {
 	async listAll(): Promise<Establishment[]> {
 		return await prisma.establishment.findMany({
+			where: {
+				deleted_at: null
+			},
 			orderBy: {
 				created_at: "desc"
 			}
@@ -12,13 +15,20 @@ export class EstablishmentPrismaRepository implements IEstablishmentRepository {
 	}
 
 	async count(): Promise<number> {
-		return await prisma.establishment.count();
+		return await prisma.establishment.count({
+			where: {
+				deleted_at: null
+			}
+		});
 	}
 
 	async paginate(page: number, limit: number): Promise<Establishment[]> {
 		return await prisma.establishment.findMany({
 			skip: (page - 1) * limit,
 			take: limit,
+			where: {
+				deleted_at: null
+			},
 			orderBy: {
 				created_at: "desc"
 			}
@@ -49,10 +59,21 @@ export class EstablishmentPrismaRepository implements IEstablishmentRepository {
 		});
 	}
 
-	async delete(id: string): Promise<Establishment> {
-		return await prisma.establishment.delete({
+	async delete(id: string, force: boolean = false): Promise<Establishment> {
+		if (force) {
+			return await prisma.establishment.delete({
+				where: {
+					id
+				}
+			});
+		}
+
+		return await prisma.establishment.update({
 			where: {
 				id
+			},
+			data: {
+				deleted_at: new Date()
 			}
 		});
 	}
